All templates now support offline mode
Add 'assets: local' to the YAML front matter of your page to work
temporarily offline. Your page will reference assets stored in
'/static'. The files in here will almost certainly be out of date, so
don't rely on this - if anyone can come up with a more robust way of
doing this...